FIX: Big refactor, create extraTokensCrossChecker comp + config.toml
